pyproject.toml
[tool.poetry]
name = "convertext"
version = "0.1.0"
description = "Lightweight universal text converter"
authors = ["dcrsn"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/dcrsn/convertext"
repository = "https://github.com/dcrsn/convertext"
keywords = ["conversion", "ebook", "pdf", "epub", "mobi", "docx", "document"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Text Processing :: Filters",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

[tool.poetry.dependencies]
python = "^3.13"
click = "^8.3.0"
pyyaml = "^6.0.3"
pypdf = "^6.1.1"
python-docx = "^1.2.0"
ebooklib = "^0.19"
odfpy = "^1.4.1"
markdown = "^3.9"
pillow = "^11.3.0"
lxml = "^6.0.2"
beautifulsoup4 = "^4.14.2"

[tool.poetry.group.mobi]
optional = true

[tool.poetry.group.mobi.dependencies]
mobi = "^0.3.3"

[tool.poetry.group.rtf]
optional = true

[tool.poetry.group.rtf.dependencies]
striprtf = "^0.0.26"

[tool.poetry.group.comics]
optional = true

[tool.poetry.group.comics.dependencies]
rarfile = "^4.2"
py7zr = "^0.22"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^6.0"
black = "^24.10"
ruff = "^0.8"
mypy = "^1.14"

[tool.poetry.scripts]
convertext = "convertext.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
